# go/model_scan.go
// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

/*
 * Du Map
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 0.0.10
 */

package openapi


import (
	"time"
)



type Scan struct {

	Time time.Time `json:"time"`

	Ores map[string]interface{} `json:"ores"`

	TileId float32 `json:"tileId"`

	Planet string `json:"planet"`

	Owner string `json:"owner"`
}

// AssertScanRequired checks if the required fields are not zero-ed
func AssertScanRequired(obj Scan) error {
	elements := map[string]interface{}{
		"time": obj.Time,
		"ores": obj.Ores,
		"tileId": obj.TileId,
		"planet": obj.Planet,
		"owner": obj.Owner,
	}
	for name, el := range elements {
		if isZero := IsZeroValue(el); isZero {
			return &RequiredError{Field: name}
		}
	}

	return nil
}

// AssertScanConstraints checks if the values respects the defined constraints
func AssertScanConstraints(obj Scan) error {
	return nil
}